+++ plugins/multi-chain-weighted-validator/toMultiChainWeightedValidatorPlugin.ts
@@ -1,4 +1,5 @@
 import type { GetKernelVersion, KernelValidator } from "@zerodev/sdk/types"
+import type { WebAuthnKey } from "@zerodev/webauthn-key"
 import type { TypedData } from "abitype"
 import {
     type UserOperation,
@@ -31,8 +32,7 @@ import {
     decodeSignatures,
     encodeSignatures
 } from "./index.js"
-import type { WebAuthnKey } from "./signers/toWebAuthnSigner.js"
-import { encodeWebAuthnPubKey } from "./signers/webAuthnUtils.js"
+import { encodeWebAuthnPubKey } from "./signers/toWebAuthnSigner.js"
 
 export type WeightedSigner = {
     account: LocalAccount