+++ wallet/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @zerodev/wallet
 
+## 0.1.3
+
+### Patch Changes
+
+- chore: upgrade passkey validator
+
+## 0.1.1
+
+### Patch Changes
+
+- fix: handle getCallsStatus
+
 ## 0.1.0
 
 ### Minor Changes